+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.2.9] - In progress
 ### Added
 - Helper to estimate categorical variable MLEs
+- Utilities for working with the SCIP solver
 
 ## [0.2.8] - 2018-03-12
 ### Updated